bugfixes and reactivate tests for prior s102 versions (2.0, 2.1, 2.2)
timeUncertainty should not have existed
dataOffsetCode was being erroneously deleted on reading rasters

+++ tests/s102/s102_test.py
@@ -27,9 +27,8 @@ def remove_file(pth, quiet=True):
         if not quiet:
             logging.warning(f"{pth} not found")
 
-@pytest.fixture(scope="module", params=[v3_0])  
+@pytest.fixture(scope="module", params=[v3_0, v2_2, v2_1, v2_0])
 def s102(request):
-    # , v2_2, v2_1, v2_0])
     yield request.param
 
 def h5py_string_comp(h5py_val, cmp_str):
@@ -177,12 +176,13 @@ def check_s102_optional_data(s102, s102obj):
         qual_name = "QualityOfSurvey"
     else:
         qual_name = "QualityOfBathymetryCoverage"
-    if qual_name in s102obj.keys():
-        assert f"{qual_name}.01" in s102obj[qual_name].keys()  # must have feature instance QualityOfBathymetryCoverage.01
-        assert qual_name in s102obj['Group_F'].keys()  # must have the Group_F entry
-    else:
-        with pytest.raises(KeyError):
-            s102obj['Group_F'].attrs[qual_name]
+        # older apis left an empty QualityOfSurvey object - only checking 3.0 and above
+        if qual_name in s102obj.keys():
+            assert f"{qual_name}.01" in s102obj[qual_name].keys()  # must have feature instance QualityOfBathymetryCoverage.01
+            assert qual_name in s102obj['Group_F'].keys()  # must have the Group_F entry
+        else:
+            with pytest.raises(KeyError):
+                s102obj['Group_F'].attrs[qual_name]
 
 
 def test_make_from_gdal(s102, bagname, output_path):
@@ -211,17 +211,18 @@ def test_make_from_bag(s102, bagname, output_path):
 
 
 def test_no_uncertainty(s102, bagname, output_path):
-    remove_file(output_path)
-    # the sample data is in NAD83 so does not meet spec - test that it's caught
-    pytest.raises(s102.S102Exception, s102.from_bag, *(bagname, output_path))
+    if s102.api.EDITION >= 3.0:
+        remove_file(output_path)
+        # the sample data is in NAD83 so does not meet spec - test that it's caught
+        pytest.raises(s102.S102Exception, s102.from_bag, *(bagname, output_path))
 
-    # override the metadata for the datum to WGS84 zone 10N and go from there
-    metadata = {"horizontalDatumReference": "EPSG", "horizontalDatumValue": 32610}
-    new_s102 = s102.from_bag(bagname, output_path, metadata=metadata)
-    assert new_s102['Group_F']['BathymetryCoverage'].shape[0] == 2
-    del new_s102.root.bathymetry_coverage.bathymetry_coverage[0].bathymetry_group[0].values.uncertainty
-    new_s102.write()
-    assert new_s102['Group_F']['BathymetryCoverage'].shape[0] == 1
+        # override the metadata for the datum to WGS84 zone 10N and go from there
+        metadata = {"horizontalDatumReference": "EPSG", "horizontalDatumValue": 32610}
+        new_s102 = s102.from_bag(bagname, output_path, metadata=metadata)
+        assert new_s102['Group_F']['BathymetryCoverage'].shape[0] == 2
+        del new_s102.root.bathymetry_coverage.bathymetry_coverage[0].bathymetry_group[0].values.uncertainty
+        new_s102.write()
+        assert new_s102['Group_F']['BathymetryCoverage'].shape[0] == 1
 
 
 def test_read_s102(s102, output_path):
